Remove the whole derevo thing. Too much trouble. Also experimentation with endpoints encodings.
When things happen I still need to add and implement manually anyway on the companion.

Doing that versus adding a @anotation and
checking which libs to import if they
exist at all that wrap what I want, and
weird errors that happen because of the
abstractions, etc...

MUCH MORE SIMPLE TO DO JUST THE USUAL THING AND BE EXPLICIT!

The derivations are always in the same place
in and expected and uniform way and can be
customized in place however one wants.